+++ crates/swc_plugin_runner/tests/css_rkyv.rs
@@ -107,6 +107,7 @@ fn invoke(input: PathBuf) -> Result<(), Error> {
             )),
             Box::new(PLUGIN_BYTES.clone()),
             Some(json!({ "pluginConfig": "testValue" })),
+            None,
         );
 
         info!("Created transform executor");
@@ -158,6 +159,7 @@ fn invoke(input: PathBuf) -> Result<(), Error> {
             )),
             Box::new(PLUGIN_BYTES.clone()),
             Some(json!({ "pluginConfig": "testValue" })),
+            None,
         );
 
         serialized_program = plugin_transform_executor
@@ -175,6 +177,7 @@ fn invoke(input: PathBuf) -> Result<(), Error> {
             )),
             Box::new(PLUGIN_BYTES.clone()),
             Some(json!({ "pluginConfig": "testValue" })),
+            None,
         );
 
         serialized_program = plugin_transform_executor